new Stats class
- store str, con and dex
- moved def out of Destructible class
- need to initialize Stats in monster generator in ActorFactory

+++ src/ai.cpp
@@ -13,7 +13,7 @@
 #include <math.h>
 using namespace std;
 
-static inline float round(float val) {
+static inline double round(double val) {
 	return floor(val + 0.5);
 }
